+++ lib/src/core/my_app.dart
@@ -1,64 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:sepp_note/src/screens/screen%201/components/home_app_bar.dart';
-import 'package:sepp_note/src/screens/screen%201/home_page_provider.dart';
+import 'package:sepp_note/src/screens/screen%201/home_page.dart';
 
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
-      create: (_) => HomePageProvider(),
-      child: const HomePage(),
-    );
-  }
-}
-
-class HomePage extends StatelessWidget {
-  const HomePage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: Scaffold(
-          body: Column(
-            children: [
-              const SizedBox(height: 50),
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 10),
-                child: Container(
-                  decoration: const BoxDecoration(
-                    color: Colors.black12,
-                    borderRadius: BorderRadius.all(Radius.circular(30)),
-                  ),
-                  height: 50,
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 20.0),
-                    child: Row(
-                      children: const [
-                        Icon(Icons.menu_outlined),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          bottomNavigationBar: const HomeBottomAppBar(),
-          floatingActionButton: FloatingActionButton(
-            backgroundColor: Colors.white,
-            shape: const RoundedRectangleBorder(
-              borderRadius: BorderRadius.all(
-                Radius.circular(15),
-              ),
-            ),
-            onPressed: () {},
-            tooltip: 'Increment Counter',
-            child: const Icon(Icons.add, color: Colors.orange, size: 45),
-          ),
-          floatingActionButtonLocation: FloatingActionButtonLocation.endDocked),
-    );
-  }
+MaterialApp myApp() {
+  return const MaterialApp(
+    debugShowCheckedModeBanner: false,
+    home: HomePage(),
+  );
 }
